+++ km_api/know_me/tests/serializers/test_profile_item_serializer.py
@@ -1,23 +1,23 @@
 from know_me import serializers
+from know_me.serializers.profile_item_content_serializers import (
+    ImageContentSerializer
+)
 
 
-def test_create(
-        media_resource_factory,
-        profile_topic_factory,
-        serializer_context):
+def test_create(profile_topic_factory, serializer_context):
     """
     Saving a serializer with valid data should create a new profile
     item.
     """
     topic = profile_topic_factory()
-    media_resource = media_resource_factory(profile=topic.group.profile)
 
     serializer_context['profile'] = topic.group.profile
 
     data = {
-        'media_resource': media_resource.pk,
         'name': 'My Profile Item',
-        'text': 'Some sample text.',
+        'image_content': {
+            'description': 'Sample image content description.',
+        }
     }
 
     serializer = serializers.ProfileItemSerializer(
@@ -28,54 +28,30 @@ def test_create(
     item = serializer.save(topic=topic)
 
     assert item.name == data['name']
-    assert item.text == data['text']
-
-    assert item.media_resource == media_resource
     assert item.topic == topic
 
+    content = item.image_content
 
-def test_create_other_user_media_resource(
-        media_resource_factory,
-        profile_factory,
-        serializer_context):
-    """
-    Users should not be able to attach a media resource from a different
-    profile to their profile item.
-    """
-    media_resource = media_resource_factory()
-    profile = profile_factory()
-
-    serializer_context['profile'] = profile
-
-    data = {
-        'media_resource': media_resource.pk,
-        'name': 'My Profile Item',
-        'text': 'I tried to attach a media resource from another profile.',
-    }
-
-    serializer = serializers.ProfileItemSerializer(
-        context=serializer_context,
-        data=data)
-
-    assert not serializer.is_valid()
+    assert content.description == data['image_content']['description']
 
 
-def test_serialize(
+def test_serialize_image_item(
         api_rf,
-        media_resource_factory,
+        image_content_factory,
         profile_item_factory,
         serializer_context):
     """
     Test serializing a profile item.
     """
-    media_resource = media_resource_factory()
-    item = profile_item_factory(media_resource=media_resource)
+    item = profile_item_factory()
+    image_content_factory(profile_item=item)
 
     serializer = serializers.ProfileItemSerializer(
         item,
         context=serializer_context)
-    media_resource_serializer = serializers.MediaResourceSerializer(
-        media_resource,
+
+    image_content_serializer = ImageContentSerializer(
+        item.image_content,
         context=serializer_context)
 
     url_request = api_rf.get(item.get_absolute_url())
@@ -84,9 +60,7 @@ def test_serialize(
         'id': item.id,
         'url': url_request.build_absolute_uri(),
         'name': item.name,
-        'text': item.text,
-        'media_resource': media_resource.pk,
-        'media_resource_info': media_resource_serializer.data,
+        'image_content': image_content_serializer.data,
     }
 
     assert serializer.data == expected
@@ -113,3 +87,30 @@ def test_update(profile_item_factory, serializer_context):
     item.refresh_from_db()
 
     assert item.name == data['name']
+
+
+def test_update_image_content(image_content_factory, serializer_context):
+    """
+    Updating a profile item's image content with nested data should work
+    properly.
+    """
+    content = image_content_factory(description='Old description.')
+    item = content.profile_item
+
+    data = {
+        'image_content': {
+            'description': 'New description.',
+        },
+    }
+
+    serializer = serializers.ProfileItemSerializer(
+        item,
+        context=serializer_context,
+        data=data,
+        partial=True)
+    assert serializer.is_valid()
+
+    serializer.save()
+    content.refresh_from_db()
+
+    assert content.description == data['image_content']['description']